+++ src/main.tsx
@@ -1,7 +1,14 @@
 import ReactDom from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
 import router from "./windows/Router";
+import { ThemeProvider } from "styled-components";
+import { GlobalStyle, LightTheme } from "shareduck-ui";
+import Header from "@components/template/Header";
 
 ReactDom.createRoot(document.getElementById("root") as HTMLElement)!.render(
-  <RouterProvider router={router} />
+  <ThemeProvider theme={LightTheme}>
+    <GlobalStyle />
+    <Header />
+    <RouterProvider router={router} />
+  </ThemeProvider>
 );
