+++ src/windows/Router.tsx
@@ -1,7 +1,6 @@
 import App from "../App";
 import { createHashRouter, RouteObject } from "react-router-dom";
 import Home from "@pages/Home";
-import Error from "./Error";
 import Category from "@pages/Category";
 import Post from "@pages/Post";
 import Overview from "@pages/Overview";
@@ -12,24 +11,26 @@ import Signup from "@pages/Signup";
 
 //todo: 서버에서 가져온 정보로 동적 라우터 생성하기
 //todo: private 라우터 설정하기
-async function loadUser() {
-  const userInfo = await window.shareDuck.invoke("user-get-ipc");
-  console.log(userInfo);
-  return false;
+function loadUser() {
+  const userInfo = window.shareDuck.invoke("user-get-ipc");
+  return userInfo;
 }
 const routers: RouteObject[] = [
+  { path: "/login", element: <Login /> },
+  { path: "/signup", element: <Signup /> },
   {
-    path: "",
+    path: "/",
     element: <App />,
-    errorElement: <Login />,
     loader: loadUser,
     children: [
       {
         index: true,
+
+        errorElement: <Login />,
         element: <Home />,
       },
-      { path: "/login", element: <Login /> },
-      { path: "/signup", element: <Signup /> },
+
+      //? 뒤에 page를 붙인 이유가 궁금합니다.
       { path: "/writepage", element: <CreatePage /> },
       { path: "/:postId/detailpage", element: <DetailPage /> },
       {
